requirements.txt
wikipedia
ddgs  # Updated from duckduckgo-search to ddgs
flask
python-dotenv
requests
langchain
langchain-community
beautifulsoup4